refactor and rename of planning table components
