+++ src/components/Planning/index.ts
@@ -1,2 +1,4 @@
-export { default as PlanningTable } from './PlanningTable';
 export { default as PlanningToolbar } from './PlanningToolbar';
+export { default as PlanningInfoPanel } from './PlanningInfoPanel';
+export { default as PlanningYearsTable } from './PlanningYearsTable';
+export { default as PlanningTable } from './PlanningTable';
